Changes that make this image bit more robust... see more.
* Label images with /image so that we can detect which image we are
in and modify behavior based on that, such that on the pages image we
don't install apt or gem depends, you wanted a Github like image.

* Fallback to apt and then bundle/gem if something fails when
installing a gem.  Instead of letting everything fail and bail
start detecting a basic error, try to apt install and try
one last time, and then bail if that fails.

* Allow users to send in $BUNDLE_ARGS so they can control things
like concurrency, they still cannot modify the path since we do
that last but that might be something for the future.

* Allow people to do $NOISY_INSTALL so that they can see all
the output that is happening.  This is akin to a debug mode for
our gem installer.  By default we are quiet but if an error
keeps happening and you can't figure out why you can
enable $NOISY_INSTALL and see all the output.

* Stop stripping end and new lines from the Gemfile when
doing an update, allow users to keep their groups and their
blank lines, we only care about our own dependencies.

* Split up the logic into methods so they are easier to
manager and cleaner to read, when all the logic was grouped
in a chaotic way, it was hard to tell what from what.