Fix #35, try to discover the UID of the user coming.
* Detect the UID of the /srv/jekyll folder and modify the Jekyll user to be that
UID so that when we run stuff it will show up as their user locally when running
in development mode.

* Modify the chown behavior to only chown vendor/bundle and _site which are the
only folders we technically need to care about, the rest are really on you but
since we detect UID even if we don't own them it shouldn't be a problem.