Travis-CI build stages aren't ready.
They are too buggy, and lack many of the desired features that Travis currently has built in.  Like Matrixes across stages, it also lacks the ability to do matrixes at all for any stage and it seems to always double up on everything for the first test stage.  Not to mention you can't install a single time before all stages.  I guess it's the nature of isolation for now?