Fix #165
We need to document to minimal users that they need to provide their own
dependencies when buliding with `jekyll/minimal`.

This also creates a bundle config change that tells Nokogiri to use
system libraries since it seems that Nokogiri's version of libxml2 is
acting a bit weird right about now.