Fix #183: Don't block plugins on pages.
It seems that Github will still try and load plugins and by blocking the
loading of plugins or even the installation of them will break the image
ultimately.  We should just allow it to happen, but without access to
any of the dependencies to install any C gems which will hopefully curb
the usage of Github Pages stuff with plugins, because you should instead
opt to use jekyll/jekyll:latest if you wish to have plugins.