Github Actions: Run deployment once a month.
Run ci periodically even if there have been no changes to the
repository. This should pick up updates to dependencies and
detect any bit rot.

Some of the starter workflows use `$cron-daily` as an example,
but the corresponding `$cron-monthly` macro doesn't seem to
be available, so I picked a random date and time.

Neither macro is mentioned in the documentation.